layout/base/tests/chrome/file_bug1018265.xhtml

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1018265
-->
<window title="Mozilla Bug 1018265"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="setTimeout(run, 0);">

  <!-- test code goes here -->
  <script type="application/javascript">
  <![CDATA[

  /** Test for Bug 1018265 **/

  var testcontent = null;

  function run() {
    testcontent = document.getElementById("testcontent");
    shouldHaveTwoNonHiddenDocumentViewers();
    testcontent.setAttribute("src", "foobarpage");
    setTimeout(errorPageLoaded, 2500)
  }

  function errorPageLoaded() {
    testcontent.addEventListener("pageshow", didGoBack, true);
    setTimeout(function() {testcontent.contentWindow.history.back();}, 0);
  }

  function didGoBack(e) {
    testcontent.removeEventListener("pageshow", didGoBack, true);
    shouldHaveTwoNonHiddenDocumentViewers();
    window.arguments[0].done();
    window.close();
  }

  function getDocumentViewer(win) {
    return win.docShell.docViewer;
  }

  function shouldHaveTwoNonHiddenDocumentViewers() {
    window.arguments[0].is(getDocumentViewer(testcontent.contentWindow).isHidden, false, "Top level DocumentViewer should not be hidden.");
    window.arguments[0].is(getDocumentViewer(testcontent.contentWindow.frames[0]).isHidden, false, " Iframe's DocumentViewer should not be hidden.");
  }
  ]]>
  </script>

  <browser type="content" id="testcontent" flex="1" src="data:text/html,&lt;iframe&gt;&lt;/iframe&gt;"/>
</window>
